Adds additional attribute collector properties.

+++ hyperion-core/src/main/java/com/willowtreeapps/hyperion/core/internal/AttributeTranslatorImpl.java
@@ -24,7 +24,7 @@ class AttributeTranslatorImpl implements AttributeTranslator {
         return measurementHelper.toDp(px) + " dp, " + px + " px";
     }
 
-    public String translatePxToSp(float px) {
+    public String translatePxToSp(int px) {
         return measurementHelper.toSp(px) + " sp, " + px + " px";
     }
